requirements.txt
librosa==0.10.0
matplotlib==3.7.1
numpy==1.22.3
opencv_python==4.6.0.66
opencv_python_headless==4.5.5.64
Pillow==8.4.0
torch==1.11.0
transformers==4.26.1
